Get record number from sequence byte array index
This implements functionality to get the associated record number at any
index within the sequence byte array. If using a sequence byte array
index to define a k-mer, this allows you to quickly check which sequence
record it is associated with.

To implement, I needed to redefine _revcomp_sba_seq_starts to be the
smaller of the two sba indices defining each chromosome's range sorted
from smallest to largest. This makes it possible to use the bisect
method for quick lookup and does not impact any implementation other
than a few unit tests that I have amended.

The implementation uses binary search, but it is possible to speed it up
to O(1) at the cost of increased RAM usage if necessary.  See the NOTE
in _get_record_num_from_sba_index for details.