Add record_names for each strand and fix naming
The main code change was to split record_names into forward and
revcomp. Without this index arithmetic was required that complicated
implementation. Given the main use case is with genomes, this results
in a small relative portion of the total memory required and simplifies
implementation.

The nomenclature is now documented in the docstring of
SequenceCollection to help keep member variables and descriptions
unambiguous. One issue was that the member _forward_sba_seq_starts
really meant segment starts in the revcomp case, which was confusing.