Update profiling organization and add fasta init
Update profiling organization and add fasta init

The organization of the run_profiling.py script needed to be modified
to accommodate the addition of fasta init profiling and other future
profiling functions. As part of this, the --small, --medium, and
--large flags were deprecated in favor of a single argument for
run_size.  It has also been updated to optionally write profiling
summaries to file.