Remove unhelpful sba indices test cases
test_get_opposite_strand_sba_indices_errors() tested whether an
exception was raised when an input index was less than zero. However,
since the dtype of the input NumPy array is uint32, it is not possible
to have a negative number, and -1 merely underflows to 2**32 - 1.  So
this test case will be removed since it does not test what was intended.