Add kmer comparison function
This function will be a fundamental operation that backs the less than
function used during quicksort of kmer_sba_start_indices and also kmer
generation and uniqueness checking. Since it is meant to be used many
times on very large arrays, it is important that it works with numba
jit.

This commit also changes the default values of break_ties in
get_is_less_than_func. When implementing comparison functionality, I
realized that breaking ties is a bad idea for sorting large arrays
since it causes a lot of unnecessary swapping.