Add generator to yield kmers by group
This function forms the basis for all kmer generator functionality. It
is meant to be a highly customizable generator that yields only the
minimum required information for downstream generators.  It is designed
to be highly customizable so that it can be reused for a variety of
applications going forward. It allows for custom kmer comparison, kmer
filtering, and group count constraints. It needs to work with numba.jit
so that it is performant for large arrays.

The only missing functionality that is anticipated to be needed in the
future is working when the kmer_sba_start_indices is defined seperately
for the forward and reverse complement strands. This requires a
modification to walk through the two arrays simultaneously and will
require an additional yield (list[bool] indicating forward /
reverse_complement for each kmer_num). This implementation has been
postponed until track_strands_separately == "both" is implemented.