Create numba funcs to get record loc info
The idea is to move functionality largely implemented in class methods
to numba @jit functions that can be used by the Kmers class @jit
functions. These functions are required to get the chromosome and seq
start index given an sba_idx. Changes have been made to the
SequenceCollection member functions to call these functions when
possible.

bisect_right() is just a lightly modified @jit version of the standard
built-in bisect module function since numba cannot determine the type
of the built-in function otherwise.