Add function to generate group size histogram
This function will back the implementation of kmers.get_kmer_count() and
kmers.get_kmer_group_counts(). It uses the kmer_info_by_group_generator
to iterate over kmers with kmer filters and group filters. Using a @jit
function rather than a pure python class method was chosen to speed up
the calculation, which takes about half the time to build the histogram.
Just counting the kmers is only marginally faster than pure python, but
it is simplest to keep it all within the same function. Interestingly,
counting the kmers takes about the same amount of time as counting the
kmers AND building the histogram. Using get_kmer_info_group_size_only is
only a very little bit faster (~5%), but is implemented since it was
simple.