Update documentation
The documentation needed to be updated to make it easier for someone new
to the genome-kmers package to get started. An algorithms page has been
added to explain at a high level how the code works. Example
calculations for the human genome have been added with links to Jupyer
notebooks. The Overview page now includes a "Getting Started" section
before review all the Basic Usage.